@@ -1,9 +1,5 @@
 import { createAsyncThunk } from '@reduxjs/toolkit';
 import { dialogsAPI } from '../../../api/dialogsAPI';
-import { BaseThunkType } from '../../store';
-// import { dialogsActionCreators, dialogsActions } from "./dialogsActions";
-
-// type ThunkType = BaseThunkType<ActionsTypes>;
 
 export const requestDialogs = createAsyncThunk(
   'dialogs/requestDialogs',
@@ -16,12 +12,15 @@ export const requestDialogs = createAsyncThunk(
 
 export const getMessages = createAsyncThunk(
   'dialogs/getMessages',
-  async (
-    { userId, isNewDialog = false }: { userId: number; isNewDialog: boolean },
-    thunkAPI
-  ) => {
+  async ({
+    userId,
+    isNewDialog = false,
+  }: {
+    userId: number;
+    isNewDialog: boolean;
+  }) => {
     if (isNewDialog) {
-      const data = await dialogsAPI.startDialog(userId);
+      await dialogsAPI.startDialog(userId);
     }
 
     const resData = await dialogsAPI.getMessages(userId);
@@ -31,10 +30,7 @@ export const getMessages = createAsyncThunk(
 
 export const sendMessage = createAsyncThunk(
   'dialogs/sendMessage',
-  async (
-    { userId, message }: { userId: number; message: string },
-    thunkAPI
-  ) => {
+  async ({ userId, message }: { userId: number; message: string }) => {
     const response = await dialogsAPI.sendMessage(userId, message);
 
     return response.data.message;
@@ -44,7 +40,7 @@ export const sendMessage = createAsyncThunk(
 export const deleteMessage = createAsyncThunk(
   'dialogs/deleteMessage',
   async (messageId: string) => {
-    const response = await dialogsAPI.deleteMessageForOwner(messageId);
+    await dialogsAPI.deleteMessageForOwner(messageId);
 
     return messageId;
   }
@@ -52,7 +48,7 @@ export const deleteMessage = createAsyncThunk(
 
 export const restoreMessage = createAsyncThunk(
   'dialogs/restoreMessage',
-  async (messageId: string, thunkAPI) => {
+  async (messageId: string) => {
     const response = await dialogsAPI.restoreMessage(messageId);
 
     return response.data.message;
@@ -61,8 +57,8 @@ export const restoreMessage = createAsyncThunk(
 
 export const addMessageToSpam = createAsyncThunk(
   'dialogs/addMessageToSpam',
-  async (messageId: string, thunkAPI) => {
-    const response = await dialogsAPI.addMessageToSpam(messageId);
+  async (messageId: string) => {
+    await dialogsAPI.addMessageToSpam(messageId);
 
     return messageId;
   }
